F OpenNebula/engineering#332: add guide on IONOS for Hosted Cloud - Ready Certification Program (#149)
**Description**

The guide is created under the solutions for the hosted cloud providers.
I have created a new folder for the images, I think we could create a
sibling folder for all future guides.

The common_101 GUI guide could be lifted out, as it could be shared
between other providers' guide.

Signed-off-by: Balazs Nemeth <bnemeth@opennebula.io>